# api/docs/swagger.yaml
definitions:
  auth_service.AddClientRequest:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      login_strategy:
        type: integer
    type: object
...
  auth_service.Client:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      login_strategy:
        type: integer
      project_id:
        type: string
    type: object
  auth_service.ClientPlatform:
    properties:
      id:
        type: string
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
...
  auth_service.ClientPrimaryKey:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
    type: object
...
  auth_service.CreateClientPlatformRequest:
    properties:
      name:
        type: string
      project_id:
        type: string
      subdomain:
        type: string
    type: object
...
  auth_service.GetClientPlatformListResponse:
    properties:
      client_platforms:
        items:
          $ref: '#/definitions/auth_service.ClientPlatform'
        type: array
      count:
        type: integer
    type: object
...
  auth_service.LogoutRequest:
    properties:
      access_token:
        type: string
    type: object
...
  auth_service.UpdateClientPlatformRequest:
    properties:
      id:
        type: string
      name:
        type: string
      subdomain:
        type: string
    type: object
  auth_service.UpdateClientRequest:
    properties:
      client_platform_id:
        type: string
      client_type_id:
        type: string
      login_strategy:
        type: integer
    type: object
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://udevs.io
paths:
  /client:
    delete:
      consumes:
      - application/json
      description: Get Client
      operationId: remove_client
      parameters:
      - description: RemoveClientBody
        in: body
        name: remove-client
        required: true
        schema:
          $ref: '#/definitions/auth_service.ClientPrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete Client
      tags:
      - Client
    post:
      consumes:
      - application/json
      description: Create Client
      operationId: create_client
      parameters:
      - description: AddClientRequestBody
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Client'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Client
      tags:
      - Client
    put:
      consumes:
      - application/json
      description: Update Client
      operationId: update_client
      parameters:
      - description: UpdateClientRequestBody
        in: body
        name: client
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Client data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Client'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Client
      tags:
      - Client
  /client-platform:
    get:
      consumes:
      - application/json
      description: Get ClientPlatform List
      operationId: get_client_platform_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetClientPlatformListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetClientPlatformListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ClientPlatform List
      tags:
      - ClientPlatform
    post:
      consumes:
      - application/json
      description: Create ClientPlatform
      operationId: create_client_platform
      parameters:
      - description: CreateClientPlatformRequestBody
        in: body
        name: client-platform
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateClientPlatformRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ClientPlatform data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.ClientPlatform'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create ClientPlatform
      tags:
      - ClientPlatform
    put:
      consumes:
      - application/json
      description: Update ClientPlatform
      operationId: update_client_platform
      parameters:
      - description: UpdateClientPlatformRequestBody
        in: body
        name: client-platform
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateClientPlatformRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ClientPlatform data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.ClientPlatform'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update ClientPlatform
      tags:
      - ClientPlatform
...
  /client-type:
    get:
      consumes:
      - application/json
      description: Get ClientType List
      operationId: get_client_type_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetClientTypeListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetClientTypeListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get ClientType List
      tags:
      - ClientType
    post:
      consumes:
      - application/json
      description: Create ClientType
      operationId: create_client_type
      parameters:
      - description: CreateClientTypeRequestBody
        in: body
        name: client-type
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreateClientTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: ClientType data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.ClientType'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create ClientType
      tags:
      - ClientType
    put:
      consumes:
      - application/json
      description: Update ClientType
      operationId: update_client_type
      parameters:
      - description: UpdateClientTypeRequestBody
        in: body
        name: client-type
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateClientTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ClientType data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.ClientType'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update ClientType
      tags:
      - ClientType
...
  /login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: login
      parameters:
      - description: LoginRequestBody
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/auth_service.LoginRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.LoginResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Login
      tags:
      - Session
  /logout:
    delete:
      consumes:
      - application/json
      description: Logout User
      operationId: logout
      parameters:
      - description: LogoutRequest
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/auth_service.LogoutRequest'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Logout User
      tags:
      - Session
  /permission:
    get:
      consumes:
      - application/json
      description: Get Permission List
      operationId: get_permission_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: search
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetPermissionListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetPermissionListResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Permission List
      tags:
      - Permission
    post:
      consumes:
      - application/json
      description: Create Permission
      operationId: create_permission
      parameters:
      - description: CreatePermissionRequestBody
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/auth_service.CreatePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Permission data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Permission'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Permission
      tags:
      - Permission
    put:
      consumes:
      - application/json
      description: Update Permission
      operationId: update_permission
      parameters:
      - description: UpdatePermissionRequestBody
        in: body
        name: permission
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdatePermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Permission data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Permission'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Permission
      tags:
      - Permission
  /permission-scope:
    delete:
      consumes:
      - application/json
      description: Get PermissionScope
      operationId: delete_permission_scope
      parameters:
      - description: PermissionScopePrimaryKeyBody
        in: body
        name: permission-scope
        required: true
        schema:
          $ref: '#/definitions/auth_service.PermissionScopePrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete PermissionScope
      tags:
      - PermissionScope
    post:
      consumes:
      - application/json
      description: Create PermissionScope
      operationId: add_permission_scope
      parameters:
      - description: AddPermissionScopeRequestBody
        in: body
        name: permission-scope
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddPermissionScopeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: PermissionScope data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.PermissionScope'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create PermissionScope
      tags:
      - PermissionScope
...
  /refresh:
    put:
      consumes:
      - application/json
      description: Refresh Token
      operationId: refresh
      parameters:
      - description: RefreshTokenRequestBody
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/auth_service.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.User'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Refresh Token
      tags:
      - Session
  /relation:
    post:
      consumes:
      - application/json
      description: Create Relation
      operationId: create_relation
      parameters:
      - description: AddRelationRequestBody
        in: body
        name: relation
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddRelationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Relation data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Relation'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Relation
      tags:
      - Relation
    put:
      consumes:
      - application/json
      description: Update Relation
      operationId: update_relation
      parameters:
      - description: UpdateRelationRequestBody
        in: body
        name: relation
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateRelationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Relation data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Relation'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Relation
      tags:
      - Relation
...
  /role:
    get:
      consumes:
      - application/json
      description: Get Roles List
      operationId: get_roles_list
      parameters:
      - description: offset
        in: query
        name: offset
        type: integer
      - description: limit
        in: query
        name: limit
        type: integer
      - description: client-platform-id
        in: query
        name: client-platform-id
        type: string
      - description: client-type-id
        in: query
        name: client-type-id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: GetRolesListResponseBody
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.GetRolesResponse'
              type: object
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Get Roles List
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Create Role
      operationId: create_role
      parameters:
      - description: AddRoleRequestBody
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create Role
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update Role
      operationId: update_role
      parameters:
      - description: UpdateRoleRequestBody
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Update Role
      tags:
      - Role
  /role-permission:
    delete:
      consumes:
      - application/json
      description: Get RolePermission
      operationId: delete_role_permission
      parameters:
      - description: RolePermissionPrimaryKeyBody
        in: body
        name: role-permission
        required: true
        schema:
          $ref: '#/definitions/auth_service.RolePermissionPrimaryKey'
      produces:
      - application/json
      responses:
        "204":
          description: ""
        "400":
          description: Invalid Argument
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Delete RolePermission
      tags:
      - RolePermission
    post:
      consumes:
      - application/json
      description: Create RolePermission
      operationId: add_role_permission
      parameters:
      - description: AddRolePermissionRequestBody
        in: body
        name: role-permission
        required: true
        schema:
          $ref: '#/definitions/auth_service.AddRolePermissionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: RolePermission data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.RolePermission'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Create RolePermission
      tags:
      - RolePermission
...
  /upsert-scope:
    post:
      consumes:
      - application/json
      description: Upsert Scope
      operationId: upsert_scope
      parameters:
      - description: UpsertScopeRequestBody
        in: body
        name: upsert-scope
        required: true
        schema:
          $ref: '#/definitions/auth_service.UpsertScopeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Role data
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/auth_service.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: Upsert Scope
      tags:
      - Scope
...
swagger: "2.0"